test: api utility refactoring - deduplication, config relocation, env alignment
