+++ src/api/config/.env.dev
@@ -1,0 +1,45 @@
+# Ambiente DEV — cart-tests
+# Copia questo file e compila i valori reali
+CHECKOUT_HOST=https://api.dev.platform.pagopa.it
+NOTICE_CODE_PREFIX=3020
+VALID_FISCAL_CODE_PA=77777777777
+
+# auth-service-tests
+FAKE_RECAPTCHA_NOT_VALIDATED=123456789AAbbCC
+AUTH_USER_EXPECT_NAME=NomeTest
+AUTH_USER_EXPECT_SURNAME=CognomeTest
+USE_BETA_BACKEND_HEADER=green
+
+# checkout-tests — payment verify
+UNKNOWN_NOTICE_CODE=302170000000000000
+UNKNOWN_FISCAL_CODE_PA=00000000001
+UNKNOWN_STAZIONE_NOTICE_CODE=002720356512737953
+UNKNOWN_STAZIONE_FISCAL_CODE_PA=77777777777
+EXPECTED_PAYMENT_REASON=TARI/TEFA 2021
+DUE_DATE=2021-07-31
+
+# checkout-tests — transactions
+AMOUNT=12000
+PSP_ID=BCITITMM
+FEE=100
+
+# checkout-tests — deployment headers
+DEPLOYMENT=green
+DEPLOYMENT_PAYMENT_METHODS=green
+DEPLOYMENT_PAYMENT_REQUESTS=green
+DEPLOYMENT_TRANSACTION_SERVICE=green
+
+# checkout-tests — operation languages
+OPERATION_LANGUAGE_IT=it
+OPERATION_LANGUAGE_FR=fr
+OPERATION_LANGUAGE_DE=de
+OPERATION_LANGUAGE_EN=en
+OPERATION_LANGUAGE_SL=sl
+
+# checkout-tests + ecommerce-cdc — NPG
+NPG_HOST=https://stg-ta.nexigroup.com
+NPG_TEST_CARD_PAN=4349940199004549
+NPG_TEST_EXPIRATION_DATE=12/99
+NPG_TEST_SECURITY_CODE=123
+NPG_TEST_CARDHOLDER_NAME=Test Test
+NPG_TEST_CARD_BRAND=VISA
