+++ src/api/utility/api_test_environment.py
@@ -2,8 +2,8 @@
 Behave environment hooks comuni a tutti i test di api-test.
 
 Responsabilità:
-- before_all : carica le variabili d'ambiente dal file config/api-tests/.env.<env>
-- before_scenario_api_test : reset dello stato condiviso tra api-test
+- before_all : carica le variabili d'ambiente dal file src/api/config/.env.<env>
+- before_scenario : reset dello stato condiviso tra api-test
 
 I moduli specifici (es. cart) possono aggiungere ulteriore logica di setup/teardown
 delegando prima a queste funzioni.
@@ -16,25 +16,19 @@ from dotenv import load_dotenv
 def _resolve_env_file(env: str) -> str:
     """
     Restituisce il percorso assoluto del file .env per l'ambiente richiesto.
-    I file sono in <repo_root>/config/api-tests/.env.<env>.
+    I file sono in <repo_root>/src/api/config/.env.<env>.
     """
-    # src/api/utility/api_test_environment.py → 4 dirname = repo root
-    repo_root = os.path.dirname(
-        os.path.dirname(
-            os.path.dirname(
-                os.path.dirname(os.path.abspath(__file__))
-            )
-        )
-    )
-    return os.path.join(repo_root, "config", "api-tests", f".env.{env}")
+    # src/api/utility/api_test_environment.py -> 2 dirname = src/api/
+    api_root = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+    return os.path.join(api_root, "config", f".env.{env}")
 
 
 def before_all(context):
     """
     Carica le variabili d'ambiente dal file .env corretto.
-    L'ambiente viene passato tramite -D env=dev|uat|local (default: dev).
+    L'ambiente viene selezionato tramite la variabile TARGET_ENV (default: dev).
     """
-    env = context.config.userdata.get("env", os.environ.get("TARGET_ENV", "dev"))
+    env = os.environ.get("TARGET_ENV", "dev")
     env_file = _resolve_env_file(env)
 
     if not os.path.exists(env_file):
